return `UNSOLVABLE` instead of `NO_PLAN_FOUND`
So, returning `NO_PLAN_FOUND` would actually be more correct, semantically. Indeed, since we operate under the assumption of a maximal number of swaps, we can't know if the problem could actually be solvable with more swaps or not. However, if we do not use `UNSOLVABLE`, the interface can't allow us to ask questions and obtain explanations about unsatisfiability.